Move Sidebar to separate component
